finish test

+++ SourceFiles/ftorch/src/kinds.F90
@@ -1,6 +1,6 @@
 module kinds
-   use iso_fortran_env, only : real64, int32
+   use iso_fortran_env, only : real64, int64
    implicit none
    public
-   integer, parameter :: rkind = real64, ikind = int32
+   integer, parameter :: rkind = real64, ikind = int64
 end module kinds
